+++ src/ape/contracts/base.py
@@ -7,12 +7,7 @@ from pydantic.dataclasses import dataclass
 
 from ape.api import Address, ReceiptAPI, TransactionAPI
 from ape.api.address import BaseAddress
-from ape.exceptions import (
-    ArgumentsLengthError,
-    ContractError,
-    ProviderNotConnectedError,
-    TransactionError,
-)
+from ape.exceptions import ArgumentsLengthError, ContractError, ProviderNotConnectedError
 from ape.logging import logger
 from ape.types import AddressType
 from ape.utils import ManagerAccessMixin, cached_property
@@ -56,11 +51,6 @@ class ContractConstructor(ManagerAccessMixin):
         )
 
     def __call__(self, *args, **kwargs) -> ReceiptAPI:
-        if "sender" in kwargs:
-            sender = kwargs["sender"]
-            txn = self.serialize_transaction(*args, **kwargs)
-            return sender.call(txn)
-
         txn = self.serialize_transaction(*args, **kwargs)
         return self.provider.send_transaction(txn)
 
@@ -180,12 +170,8 @@ class ContractTransaction(ManagerAccessMixin):
         )
 
     def __call__(self, *args, **kwargs) -> ReceiptAPI:
-        if "sender" in kwargs:
-            sender = kwargs["sender"]
-            txn = self.serialize_transaction(*args, **kwargs)
-            return sender.call(txn)
-
-        raise TransactionError(message="Must specify a `sender`.")
+        txn = self.serialize_transaction(*args, **kwargs)
+        return self.provider.send_transaction(txn)
 
 
 class ContractTransactionHandler(ManagerAccessMixin):
